TFJson: Prepend , if entry added is not the first.
The old logic always appended the , after every entry and removed the
last one when calling endArray() or endObject().

However this required one byte more than expected when the buffer size
is exactly the required one, as the top level array or object
temporarily gets a , that is removed with end().