OcppTools: Add enum aliases. Warn if lookup_key result is ignored.
This allows us to specify alias names for enum keys. We need those
to handle centrals that for example don't handle the
FormatViolation/FormationViolation correctly.

This also adds a warning if the result of lookup_key is ignored.
Ignoring the result leads to bugs such as the one fixed here in
OcppConnection::handleMessage.