Connector: Unpluging the cable in a transaction is not unexpected.
A charger with a fixed cable reports NotConnected if the cable
is unplugged at the vehicle. So this _can_ happen in a transaction
and is almost equivalent to the charger reporting PlugDetected:
On PlugDetected, we still have to decide whether to unlock the
plug. On NotConnected, the plug was either not sufficiently locked,
or there is no plug (i.e. the charger has a fixed cable).