tfjson: Fix out-of-bounds buffer access during number parsing
If the input buffer contains exactly one number without any whitespace
or nul-terminator following it then an out-of-bounds write and read
happened while trying to nul-terminate the input buffer in order to use
strtod/strto[u]ll to parse the number.

Add an option to the parser that allows to make a nul-terminated copy
of the number string. If the option is disabled parsing will now report
an error in this situation.