Connection: Decouple response sending from request processing.
A platform's web socket client can be locked while processing a request,
so sending the response/error immediately could not work.

Send a response/error in the next tick. We don't need a queue here,
because only a single request can be in flight. If we receive another
request while the first response was not sent yet, the first request
already timed out and we can throw away the response.