MeterValueHandler, Accumulator: Fix MeterValues timestamps and reg vals.
The spec wants the timestamp of clock aligned meter values to be the
start of the interval, but does not specify what the timestamp of
sampled meter values is. Also it is unspecified whether register values
are measured at the start or the end of an interval.

To be more consistent, always set the timestamp to the start of an
interval. Also read register values at the start, not the end so that
the reported register value belongs to the sent timestamp.